fix: resolve circular depedancy on HashConnector
